chore(page): restructure event code
